Speed up, changes and a fix.
Sped up the GUI so it feels more responsive.
(Warning, doing a large number of card sets could crash the app.)
Changed the Set selector so that it grabs a list from The Gatherer.
Changed it so that the text area show json objects of card data.
Fixed missing SetMeta.cs issue.